Move `limitedLocalTimeRange` under citizen-frontend
